+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 **Internal**:
 
-- Remove metrics summaries. ([#4278](https://github.com/getsentry/relay/pull/4278))
+- Remove metrics summaries. ([#4278](https://github.com/getsentry/relay/pull/4278), [#4279](https://github.com/getsentry/relay/pull/4279))
 
 ## 24.11.0